Move the rest of the copy code to the component
