+++ src/components/CopyToClipboardButton/styles.module.less
@@ -1,5 +1,11 @@
 @import '../styles.module.less';
 
+.container {
+  display: flex;
+  align-items: center;
+  gap: 14px;
+}
+
 .buttonStyling {
   .socialShareButton;
 
@@ -22,5 +28,13 @@
   &:hover {
     background-color: #00A99D30;
     border-color: #00A99D;
+  }
+}
+
+.copyFailed {
+  width: 100%;
+
+  .MuiInputBase-input {
+    padding: 12px;
   }
 }
