+++ src/components/styles.module.less
@@ -62,7 +62,7 @@
   justify-content: center;
   border: 1px solid #D7DCE5;
   border-radius: 100%;
-  width: 40px;
+  min-width: 40px;
   height: 40px;
   background-color: #FFFFFF;
   cursor: pointer;
